Separate out the script loader from the SQLScript class
