+++ report/sql_tasks/__init__.py
@@ -1,1 +1,3 @@
 """Classes for reading and executing tasks defined in SQL files."""
+
+from report.sql_tasks.loader import from_dir
